+++ include/absent/combinators/fmap.h
@@ -2,7 +2,7 @@
 #define RVARAGO_ABSENT_FMAP_H
 
 #include "absent/member.h"
-#include "absent/nullable.h"
+#include "absent/syntax/nullable.h"
 
 #include <functional>
 #include <utility>
@@ -21,10 +21,10 @@ namespace rvarago::absent {
      */
     template <typename A, template <typename> typename Nullable, typename Mapper>
     constexpr auto fmap(Nullable<A> const& input, Mapper fn) -> Nullable<decltype(fn(std::declval<A>()))> {
-        if (nullable::empty(input)) {
+        if (syntax::nullable::empty(input)) {
             return {};
         }
-        return fn(nullable::value(input));
+        return fn(syntax::nullable::value(input));
     }
 
     /***
